2_Orchestrator_And_ML_(Python)/requirements.txt
alpaca-trade-api-python
requests
vaderSentiment
python-dotenv
scikit-learn
joblib
pandas
matplotlib